Post refactor updates
