+++ io_scene_niftools/__init__.py
@@ -59,7 +59,7 @@ from io_scene_niftools import properties, operators, ui
 
 from io_scene_niftools.utils.util_logging import NifLog
 with open(os.path.join(current_dir, "VERSION.txt")) as version:
-    NifLog.info(f"Loading: Blender Niftools Addon: {version.read():s}")
+    NifLog.info(f"Loading: Blender Niftools Addon: {version.read()}")
 
 import pyffi
 NifLog.info(f"Loading: Pyffi: {pyffi.__version__}")
