+++ io_scene_niftools/addon_updater.py
